-added performance curves api and repository functions -added mappings between performance curve entities and dtos -updated performance curves ui to use new naming of performance curves, new api endpoints, changes with the models and database schema -added criterion library api and repository functions -added criterion library dtos and mappings between the dtos and entities -updated criterion library ui and service functions to work with new api endpoints, models, and database schema -updated equation editor ui to match model changes and database schema changes -added uuid utilities to client app -added library utilities to client app

+++ BridgeCareApp/AppliedResearchAssociates.iAM.DataPersistenceCore/Repositories/MSSQL/Extensions/IAMContextExtensions.cs
@@ -11,6 +11,9 @@ namespace AppliedResearchAssociates.iAM.DataPersistenceCore.Repositories.MSSQL.E
 {
     public static class IAMContextExtensions
     {
+        private static readonly bool IsRunningFromXUnit = AppDomain.CurrentDomain.GetAssemblies()
+            .Any(a => a.FullName.ToLowerInvariant().StartsWith("xunit"));
+
         public static void AddOrUpdate<T>(this IAMContext context, T entity, Guid key) where T : class
         {
             if (entity == null)
@@ -63,10 +66,28 @@ namespace AppliedResearchAssociates.iAM.DataPersistenceCore.Repositories.MSSQL.E
         {
             var entities = context.Set<T>();
 
-            var entitiesToDelete = entities.Where(predicate);
+            var entityToDelete = entities.SingleOrDefault(predicate);
+            if (entityToDelete != null)
+            {
+                entities.Remove(entityToDelete);
+            }
+        }
+
+        public static void DeleteAll<T>(this IAMContext context, Expression<Func<T, bool>> predicate) where T : class
+        {
+            var entities = context.Set<T>();
+
+            var entitiesToDelete = entities.Where(predicate).ToList();
             if (entitiesToDelete.Any())
             {
-                entities.RemoveRange(entitiesToDelete);
+                if (IsRunningFromXUnit)
+                {
+                    entities.RemoveRange(entitiesToDelete);
+                }
+                else
+                {
+                    context.BulkDelete(entitiesToDelete);
+                }
             }
         }
